pyproject.toml
[project]
name = "bv-rag"
version = "1.0.0"
description = "Maritime regulation voice Q&A system powered by RAG"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "python-multipart>=0.0.18",
    "websockets>=14.0",

    # AI APIs
    "openai>=1.60",
    "anthropic>=0.42",

    # Databases
    "qdrant-client>=1.12",
    "asyncpg>=0.30",
    "psycopg2-binary>=2.9",
    "redis>=5.2",

    # Crawling & Parsing
    "scrapy>=2.12",
    "beautifulsoup4>=4.12",
    "lxml>=5.3",

    # Utilities
    "tiktoken>=0.8",
    "pydantic-settings>=2.7",
    "tenacity>=9.0",
    "rich>=13.9",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = [
    "config*",
    "crawler*",
    "parser*",
    "chunker*",
    "db*",
    "retrieval*",
    "generation*",
    "voice*",
    "memory*",
    "pipeline*",
    "api*",
    "evaluation*",
    "scripts*",
]
